requirement.txt
fastapi
uvicorn
Pillow
numpy
torch
pydantic
opencv-python
ultralytics
requests
tensorflow
h5py
deepface
torchvision 



